requirements.txt
asgiref==3.10.0
Django==5.2.8
djangorestframework==3.16.1
gunicorn==21.2.0
load-dotenv==0.1.0
psycopg2==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.3
django-cors-headers==4.3.1
django-unfold==0.19.0
langchain-community
langchain-core
langchain-openai
openai
aiogram